add guest user modal styled
